requirements.txt
jieba
numpy
scipy
bert-serving-server
bert-serving-client
